Changed all tests from Midje to clojure.test.
This may seem unnecessary, but it will help us in several ways. First,
it's faster to run, which is nice. The tests are more more simple to
read as well. Lastly, this allows us to use tools like cloverage to get
code coverage stats and test.generative to generate random test data.