Remove placeholders
Placeholders seem too noncommital. They suggest a possible value, but
you're obligated to retype it in if you want to keep it. And they
can't really be applied to all fields equally (such as group by)
without swaying the premise of the query. I'd rather skip any
potential ambiguity and rely on the tooltips and autocomplete to
handle the suggestion legwork.